id값 및 배열 int로 변경 list리턴 시 정렬 기능 추가

+++ model/message.go
@@ -2,7 +2,7 @@ package model
 
 // Message 커뮤니케이션용 객체
 type Message struct {
-	ID      int64  `json:"id,omitempty"`
+	ID      int    `json:"id,omitempty"`
 	Sender  string `json:"sender"`
 	Message string `json:"message"`
 }
